[REF] *: move the `deprecated` decorator from ORM to tools
It probably shouldn't have been in the ORM if it was going to be used
from outside of it anyway, it has nothing to do with the ORM API.
