+++ addons/base_automation/models/base_automation.py
@@ -15,6 +15,7 @@ from odoo.exceptions import LockError, MissingError
 from odoo.fields import Domain
 from odoo.http import request
 from odoo.tools import safe_eval
+from odoo.tools.func import deprecated
 
 _logger = logging.getLogger(__name__)
 
@@ -1088,7 +1089,7 @@ class BaseAutomation(models.Model):
     def _get_calendar(self, automation, record):
         return automation.trg_date_calendar_id
 
-    @api.deprecated("Since 19.0, use _cron_process_time_based_automations")
+    @deprecated("Since 19.0, use _cron_process_time_based_automations")
     def _check(self, automatic=False, use_new_cursor=False):
         if not automatic:
             raise RuntimeError("can run time-based automations only in automatic mode")
