+++ odoo/api/__init__.py
@@ -7,7 +7,6 @@ from odoo.orm.decorators import (
     constrains,
     depends,
     depends_context,
-    deprecated,
     model,
     model_create_multi,
     onchange,
